requirements.txt
fastapi
uvicorn
sqlmodel
psycopg2-binary
python-multipart
kafka-python